Implement two-way selection of notes
I slightly modified the existing eventbus setup to support up to 15 notes for two octave
scales. Scales in the scales.js require a startOffset (number of notes from 1st
finger on the 4th string) so that they will be correctly aligned with the finger
patterns. Strings in the scales.js file are number 4-1 from back to front as is
conventional in music, but internally they are zero-indexed from the lowest string to
vastly simplify the calculation of which note to highlight.