Implement selection of Instrument and Scale
Instrument and Scale selection elements are accessible and functional. This introduced a bug
where the notes were not re-rendering. Adding a key property to the notes display fixed this
issue. If a scale or instrument is selected that is missing, the errors cause the entire page
to freeze and become non-responsive. This will be fixed by adding all the possible scales.